% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnma.R
\docType{data}
\name{certolizumab}
\alias{certolizumab}
\title{Trials of certolizumab pegol (CZP) for the treatment of rheumatoid arthritis in patients}
\format{A list of Outcomes, Treat, Study, N, covariate, and Treat.order}
\usage{
certolizumab
}
\description{
A dataset of 12 trials for investigating CZP for the treatment for those who had failed on disease-modifying antirheumatic drugs, including methotrexate (MTX).
Data provides the number of patients who have improved and there are 6 different treatments with placebo. Mean disease duration (years) is provided as a covariate.
}
\references{
S. Dias, A.J. Sutton, A.E. Ades, and N.J. Welton (2013b), \emph{Heterogeneity-Subgroups, Meta-Regression, Bias, and Bias-Adjustment}, Medical Decision Making 33(5):618-640. [\url{https://doi.org/10.1177/0272989X13485157}]
}
\keyword{datasets}
