% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.network.R
\name{contrast.network.deviance.plot}
\alias{contrast.network.deviance.plot}
\title{Make a contrast network deviance plot}
\usage{
contrast.network.deviance.plot(result)
}
\arguments{
\item{result}{Object created by \code{\link{contrast.network.run}} function}
}
\value{
None
}
\description{
This makes a contrasrt network deviance plot which plots residual deviance (resdev_study) vs. all study.
}
\examples{
network <- with(parkinsons_contrast, {
contrast.network.data(Outcomes, Treat, SE, na, V)
})
\donttest{
result <- contrast.network.run(network)
contrast.network.deviance.plot(result)
}
}
