% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.network.R
\name{contrast.network.leverage.plot}
\alias{contrast.network.leverage.plot}
\title{Make a leverage plot}
\usage{
contrast.network.leverage.plot(result)
}
\arguments{
\item{result}{Object created by \code{\link{contrast.network.run}} function}
}
\description{
This function makes a leverage vs. square root of residual deviance plot
}
