% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{draw.network.graph}
\alias{draw.network.graph}
\title{Draws network graph using igraph package}
\usage{
draw.network.graph(network, label.dist = 2)
}
\arguments{
\item{network}{Object created by \code{\link{network.data}} function}

\item{label.dist}{distance of the label from the node. Default is 2.}
}
\value{
None
}
\description{
This function draws network graph using igraph package
}
\examples{
#cardiovascular
network <- with(thrombolytic, {
 network.data(Outcomes, Study, Treat, N=N, response = "binomial")
})
draw.network.graph(network)
}
