% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.autocorr.plot}
\alias{network.autocorr.plot}
\title{Generate autocorrelation plot using coda package}
\usage{
network.autocorr.plot(result, extra.pars = NULL, only.pars = NULL)
}
\arguments{
\item{result}{Object created by \code{\link{network.run}} function}

\item{extra.pars}{Extra parameters that the user wants to plot other than the default parameters.}

\item{only.pars}{Parameters that user wants to display. This gets rids of other default parameters user doesn't want to show}
}
\description{
This function plots autocorrelation using coda package.
}
\examples{
#cardiovascular
Study <- cardiovascular[["Study"]]
Treat <- cardiovascular[["Treat"]]
Outcomes <- cardiovascular[["Outcomes"]]
N <- cardiovascular[["N"]]
network <- network.data(Outcomes, Study, Treat, N, response = "multinomial")
result <- network.run(network)
network.autocorr.plot(result, only.pars = "d")
}
