% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.deviance.plot}
\alias{network.deviance.plot}
\title{Make a deviance plot}
\usage{
network.deviance.plot(result)
}
\arguments{
\item{result}{Object created by \code{\link{network.run}} function}
}
\value{
None
}
\description{
This makes a deviance plot which plots residual deviance (dev_arm) vs. all the arms for each study.
}
\examples{
network <- with(blocker, {
 network.data(Outcomes, Study, Treat, N = N, response = "binomial")
})
\donttest{
result <- network.run(network)
network.deviance.plot(result)
}
}
