% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.gelman.plot}
\alias{network.gelman.plot}
\title{Use coda package to plot Gelman-Rubin diagnostic plot}
\usage{
network.gelman.plot(result, extra.pars = NULL, only.pars = NULL)
}
\arguments{
\item{result}{Object created by \code{\link{network.run}} function}

\item{extra.pars}{Extra parameters that the user wants to plot other than the default parameters.}

\item{only.pars}{Parameters that user wants to display. This gets rids of other default parameters user doesn't want to show.}
}
\description{
This function plots Gelman-Rubin diagnostic using coda package.
}
\examples{
#blocker
network <- with(blocker,{
 network.data(Outcomes, Study, Treat, N = N, response = "binomial")
})
result <- network.run(network)
network.gelman.plot(result, only.pars = "d")
}
