% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.leverage.plot}
\alias{network.leverage.plot}
\title{Make a leverage plot}
\usage{
network.leverage.plot(result)
}
\arguments{
\item{result}{Object created by \code{\link{network.run}} function}
}
\description{
This function makes a leverage vs. square root of residual deviance plot
}
