% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{network.rankogram}
\alias{network.rankogram}
\title{Create a treatment rankogram}
\usage{
network.rankogram(result, title = "Rankogram", catnames = NULL)
}
\arguments{
\item{result}{Object created by \code{\link{network.run}} function}

\item{title}{Title of the rankogram}

\item{catnames}{Label for category comparison in multinomial outcomes}
}
\description{
Another plot to display how each treatment is ranked. For each rank, we show how likely each treatment will be at that rank.
}
\examples{
network <-with(blocker, {
 network.data(Outcomes, Study, Treat, N = N, response = "binomial")
})
result <- network.run(network)
network.rankogram(result)
}
\seealso{
\code{\link{rank.tx}}
}
