% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnma.R
\docType{data}
\name{parkinsons}
\alias{parkinsons}
\title{Dopamine agonists as adjunct therapy in Parkinson's disease}
\format{
A list of Outcomes, Treat, Study, N, covariate, and Treat.order
}
\usage{
parkinsons
}
\description{
A dataset of 7 studies investigating the mean lost work-time reduction in patients given
4 dopamine agonists and placebo as adjunct therapy for Parkinson's disease.
There is placebo and four active drugs coded 2 to 5.
}
\references{
S. Dias, A.J. Sutton, A.E. Ades, and N.J. Welton (2013a), \emph{A Generalized Linear Modeling Framework for Pairwise and Network Meta-analysis of Randomized Controlled Trials}, Medical Decision Making 33(5):607-617. \doi{10.1177/0272989X12458724}
}
\keyword{datasets}
