% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnma.R
\docType{data}
\name{parkinsons_contrast}
\alias{parkinsons_contrast}
\title{Dopamine agonists as adjunct therapy in Parkinson's disease}
\format{
A list of Outcomes, Treat, SE, na, and V
}
\usage{
parkinsons_contrast
}
\description{
A contrast level (i.e. treatment difference) dataset of 7 studies investigating the mean lost work-time reduction in patients given
4 dopamine agonists and placebo as adjunct therapy for Parkinson's disease. Placebo is coded as 1, and four active drugs are coded 2 to 5.
There is placebo, coded as 1, and four active drugs coded 2 to 5.
}
\references{
S. Dias, A.J. Sutton, A.E. Ades, and N.J. Welton (2013a), \emph{A Generalized Linear Modeling Framework for Pairwise and Network Meta-analysis of Randomized Controlled Trials}, Medical Decision Making 33(5):607-617. \doi{10.1177/0272989X12458724}
}
\keyword{datasets}
