% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ume.network.R
\name{plot.ume.network.result}
\alias{plot.ume.network.result}
\title{Plot traceplot and posterior density of the result using contrast data}
\usage{
\method{plot}{ume.network.result}(x, ...)
}
\arguments{
\item{x}{Result object created by \code{\link{ume.network.run}} function}

\item{...}{Additional arguments affecting the plot produced}
}
\value{
None
}
\description{
This function uses plotting function in coda package to plot mcmc.list object
}
\examples{
network <- with(smoking, {
 ume.network.data(Outcomes, Study, Treat, N = N, response = "binomial", type = "random")
})
\donttest{
result <- ume.network.run(network)
plot(result)
}
}
