% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.data.R
\name{preprocess.data}
\alias{preprocess.data}
\title{Preprocess data given user input}
\usage{
preprocess.data(Outcomes = NULL, Study = NULL, Treat = NULL,
  N = NULL, SE = NULL, response = NULL, Treat.order = NULL,
  type = "random", baseline = "none", baseline.risk = "independent",
  covariate = NULL, covariate.type = NULL, covariate.model = NULL)
}
\description{
This is first bit of our code that runs when the user enters the data using \code{\link{network.data}}.
The function preprocesses the given data to fit the data format necessary to run our NMA model in JAGS.
It is essential to know how the data is preprocessed to obtain correct analysis of interest.
Thus, we have exported the function so that user can have better understanding on what is going on underneath.
}
