% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{rank.tx}
\alias{rank.tx}
\title{Create a treatment rank table}
\usage{
rank.tx(result)
}
\arguments{
\item{result}{Object created by \code{\link{network.run}} function}
}
\value{
Returns a table of ranking
}
\description{
This function makes a table of ranking for each treament. Each number in the cell represents a probability certain treatment was in such rank.
This table is also stored as an output from \code{\link{network.run}}.
}
\examples{
network <- with(blocker, {
 network.data(Outcomes, Study, Treat, N = N, response = "binomial")
})
\donttest{
result <- network.run(network)
rank.tx(result)
}
}
\seealso{
\code{\link{network.rank.tx.plot}}
}
