% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{relative.effects}
\alias{relative.effects}
\title{Find relative effects for base treatment and comparison treatments}
\usage{
relative.effects(result, base.treatment = NULL,
  comparison.treatments = NULL, base.category = NULL,
  comparison.categories = NULL, covariate = NULL)
}
\arguments{
\item{result}{Object created by \code{\link{network.run}} function}

\item{base.treatment}{Base treatment user wants for the relative effects. Base treatment is initially set by \code{Treat.order} parameter in \code{\link{network.data}} (first one in the list). If set to null, default is to use base treatment.}

\item{comparison.treatments}{Treatments that user wants to compare against base treatment. If set to null, all the treatments besides base treatment is considered as comparison treatments.}

\item{base.category}{Base category user wants for the relative effects. Only used for multinomial data.}

\item{comparison.categories}{Category that user wants to compare against base.category. Only used for multinomial data.}

\item{covariate}{Covariate value at which to compute relative effects. Only used if covariate value is specified in the model.}
}
\value{
This returns a mcmc.list sample of relative effects for the base treatment specified. This allows user to obtain relative effects of different base.treatment after the sampling has been done.
For a simple summary, use \code{\link{relative.effects.table}}.
}
\description{
This function calculates relative effects for base treatment and comparison treatments.
}
\examples{
#We can fit two different models with different base treatment and we can
#obtain same relative effects estimate using this function
#parkinsons
network <- with(parkinsons, {
 network.data(Outcomes, Study, Treat, SE = SE, response = "normal")
})
result <- network.run(network) 
summary(result)

network2 <- with(parkinsons, {
 network.data(Outcomes, Study, Treat, SE = SE, response = "normal",
 Treat.order = c(2,1,3,4,5))
})
result2 <- network.run(network2)

summary(result)
summary(relative.effects(result2, base.treatment = 1))

#This also works for comparing different base.category for multinomial.
#We fit two different models and compare the estimates again.
#cardiovascular

network3 <- with(cardiovascular, {
 network.data(Outcomes, Study, Treat, N, response = "multinomial")
})
result3 <- network.run(network3)

network4 <- with(cardiovascular, {
 network.data(Outcomes[,c(2,1,3)], Study, Treat, N, response = "multinomial")
})
result4 <- network.run(network4)

summary(result3)
summary(relative.effects(result4, base.category = 2))
}
\seealso{
\code{\link{relative.effects.table}}
}
