% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnma.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{Smoking cessation counseling programs}
\format{A list of Outcomes, Treat, Study, and N}
\usage{
smoking
}
\description{
Twenty-four studies, including 2 three-arm trials, compared 4 smoking cessation counseling programs and recorded the number of
individuals with successful smoking cessation at 6 to 12 month. Counseling programs include 1 = no intervention, 2 = self-help,
3 = individual counseling, and 4 - group counseling.
}
\references{
S. Dias, A.J. Sutton, A.E. Ades, and N.J. Welton (2013a), \emph{A Generalized Linear Modeling Framework for Pairwise and Network Meta-analysis of Randomized Controlled Trials}, Medical Decision Making 33(5):607-617. [\url{https://doi.org/10.1177/0272989X12458724}]
}
\keyword{datasets}
