% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnma.R
\docType{data}
\name{statins}
\alias{statins}
\title{Trials of statins for cholesterol lowering vs. placebo or usual care}
\format{
A list of Outcomes, Treat, Study, N, covariate, and Treat.order
}
\usage{
statins
}
\description{
A dataset of 19 trials of statins for cholesterol lowering vs. placebo.
Each trial has a subgroup indicator for primary prevention (patients included had no previous heart disease) or
secondary prevention (patients had previous heart disease). Dummy variable is coded such that covariate is equal to 1
if a study is a secondary prevention study and 0 if a study is a primary prevention study.
}
\references{
S. Dias, A.J. Sutton, A.E. Ades, and N.J. Welton (2013b), \emph{Heterogeneity-Subgroups, Meta-Regression, Bias, and Bias-Adjustment}, Medical Decision Making 33(5):618-640. \doi{10.1177/0272989X13485157}
}
\keyword{datasets}
