% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnma.R
\docType{data}
\name{thrombolytic}
\alias{thrombolytic}
\title{Thrombolytic drugs and percutaneous transluminal coronary angioplasty}
\format{
A list of Outcomes, Treat, Study, and N
}
\usage{
thrombolytic
}
\description{
A dataset consisting of 50 trials comparing 8 thrombolytic drugs and percutaneous transluminal coronary angioplasty, following
acute myocardial infarction. Data consist of the number of deaths in 30 or 35 days and the number of patients in each treatment arm.
There are 9 treatments in total: streptokinase (1), alteplase (2), accelerated alteplase (3), streptokinase + alteplase (4),
reteplase (5), tenecteplase (6), percutaneous transluminal coronary angioplasty (7), urokinase (8), anistreptilase (9)
}
\references{
S. Dias, A.J. Sutton, A.E. Ades, and N.J. Welton (2013a), \emph{A Generalized Linear Modeling Framework for Pairwise and Network Meta-analysis of Randomized Controlled Trials}, Medical Decision Making 33(5):607-617. \doi{10.1177/0272989X12458724}
}
\keyword{datasets}
