% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.summary.R
\name{variance.tx.effects}
\alias{variance.tx.effects}
\title{Calculate correlation matrix for multinomial heterogeneity parameter.}
\usage{
variance.tx.effects(result)
}
\arguments{
\item{result}{Object created by \code{\link{network.run}} function}
}
\value{
Returns correlation matrix
}
\description{
This function calculates correlation matrix from the variance matrix for heterogeneity parameter. Only used for multinomial.
}
\examples{
#cardiovascular
network <- with(cardiovascular, {
 network.data(Outcomes, Study, Treat, N, response = "multinomial")
})
\donttest{
result <- network.run(network)
variance.tx.effects(result)
}
}
